Make Report PDF/requirements.txt
Flask
fpdf2
pymongo
python-dotenv
gunicorn